proxy: Move Request and Response to the api package
These structures define the JSON snippets transporting payloads, so they
are really part of the proxy API.
